Example of Android Push Notification with Firebase Console
This example demonstrates receiving push notifications from the Firebase
Console, and then displays the content on the screen in JSON format. You
can send push notifications to the devices running this example by using
the notifications composer in the Firebase Console. The Firebase servers
will notify the registered application that the new message is available
to sync.

Task-number: QTPM-182
Change-Id: Ie84ce7c1446af3a6bcf7db4be9911db2b18f08ef
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Ari Salmi <snowgrains@snowgrains.com>