Example of Kaltiot Push Notification for Embedded devices
This example demonstrates receiving push notifications from the Kaltiot
IoT Cloud, and then displays the content on the screen. The Kaltiot
gateway daemon will notify the registered application that the new
message is available to sync. You can send push notifications to the
devices running this example by using the IoT-Devices service in the
Kaltiot Console or Kaltiot HTTP Rest API with curl command. This example
works on both Linux64 and Raspberry Pi devices.

Task-number: QTPM-182
Change-Id: Ifbcf3efbbcdef28e44ce46d019806510dc34194b
Reviewed-by: Ryan Chu <ryan.chu@qt.io>
Reviewed-by: Ari Salmi <snowgrains@snowgrains.com>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>