Return error status to inform the user when client registration fails
When connecting a client id to the Kaltiot-service provider fails, the
error status of Kaltiot SDK should be returned to the user application.
So I rework some functions and return the error status from the Kaltiot
library to the user application. To prevent memory leak, the new client
instance should be saved when connecting the client to a provider, and
then released when removing the client.

Change-Id: I1e6248882a400cfcca78af5f1151cc61b6c954ed
Reviewed-by: Ryan Chu <ryan.chu@qt.io>
Reviewed-by: Ari Salmi <snowgrains@snowgrains.com>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>