Create kaltiot_daemon_path option to set the path of Kaltiot daemon
When calling the ks_gw_client_connect function of Kaltiot library with
NULL path, the Kaltiot library will connect to the default Unix domain
socket at /tmp/ks_client_socket which needs root permission. The problem
is that there is no alternative option for the user to change the value
of path. To make it flexible, I create an option (kaltiot_daemon_path)
of client parameter to set the path when connecting a client id to the
Kaltiot-service provider. This option lets the Kaltiot library connect
to the port which Kaltiot daemon is listening on.

Change-Id: Ia90a8ffa8b7e4059acfad4811f8e3177f6223179
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
Reviewed-by: Ryan Chu <ryan.chu@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Ari Salmi <snowgrains@snowgrains.com>